Revert "fsd"
