+++ src/app/components/Header.tsx
@@ -1,10 +1,11 @@
 "use client"
 
-import './styles.scss';
+import './../styles.scss';
+import './../styles/components/_header.scss';
 import Image from "next/image";
+import Button from "./Button";
 import Link from "next/link";
 import React from 'react';
-import Button from '../button/Button';
 
 
 export default function Header() {
