+++ src/app/components/Slider.tsx
