+++ src/app/styles/components/_button.scss
@@ -15,14 +15,7 @@
         color: var(--button-secondary-border-color);
         font-size: 16px;
         border: 1px solid var(--color-accent);
-        color: var(--color-accent);
-        transition:
-            background-color 0.3s;
-            
-        &:hover {
-            background-color: var(--color-accent);
-            color: var(--color-light);
-        }
+        color: var(--color-accent)
     }
 
     &__cart {
@@ -33,7 +26,6 @@
         color: var(--color-light);
         border: none;
     }
-
     &__cart-alt {
         height: 40px;
         min-width: 140px;
@@ -41,10 +33,6 @@
         background-color: transparent;
         color: var(--color-accent);
         border: 1px solid var(--color-accent);
-        &:hover {
-            background-color: var(--color-accent);
-            color: var(--color-light);
-        }
     }
 
     &__location {
