+++ src/app/styles/components/_card.scss
