+++ src/app/styles/components/_category.scss
@@ -1,4 +1,4 @@
-@import '../../app/styles';
+@import './../../styles';
 
 .category {
     display: flex;
