+++ src/app/styles/components/_footer.scss
@@ -1,4 +1,4 @@
-@import '../../app/styles';
+@import './../../styles';
 
 .footer {
     padding: 80px 0px;
@@ -24,14 +24,14 @@
         @include desktop {
             display: flex;
             flex-direction: column;
-            padding: 10px 0px;
+            padding: 10px 0;
         }
     }
 
     &__logo {
         display: flex;
         flex-direction: column;
-        gap: 10px;
+        gap: 10;
 
         @include desktop {
             margin-bottom: 10px;
@@ -82,11 +82,11 @@
     }
 
     &__adress {
-        padding: 20px 0px;
+        padding: 20px 0;
     }
 
     &__copy {
-        padding: 20px 0px;
+        padding: 20px 0;
     }
 
 }
