+++ src/app/styles/components/_header.scss
@@ -1,4 +1,4 @@
-@import '../../app/styles';
+@import './../../styles';
 
 .header {
     display: flex;
