+++ src/app/styles/components/_modal.scss
@@ -1,4 +1,4 @@
-@import '../../app/styles';
+@import './../../styles';
 
 .modal {
   position: relative;
