+++ src/app/styles/components/_sets.scss
@@ -1,4 +1,4 @@
-@import '../../app/styles';
+@import './../../styles';
 
 .sets {
     display: flex;
@@ -38,16 +38,6 @@
         @include mobile {
             display: flex;
             justify-content: center;
-        }
-    }
-
-    &__more {
-        display: flex;
-        justify-content: center;
-        padding: 40px 0;
-
-        button {
-            padding: 0px 25px;
         }
     }
 }
